labs/lab2-rag-policy-bot/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
openai==1.63.2
azure-search-documents==11.6.0
